Add speed slider, new HeronLX with deltaMs as double

+++ ArjunBanker.pde
@@ -13,7 +13,7 @@ class TelevisionStatic extends SCPattern {
     addParameter(hueParameter);
   }
 
- void run(int deltaMs) {
+ void run(double deltaMs) {
     boolean d = direction.getValuef() > 5.0;
     for (Point p : model.points) {             
       colors[p.index] = color((lx.getBaseHuef() + random(hueParameter.getValuef() * 360))%360, random(saturationParameter.getValuef() * 100), random(brightParameter.getValuef() * 100));
@@ -37,7 +37,7 @@ class AbstractPainting extends SCPattern {
     img.loadPixels();    
   } 
  
-  void run(int deltaMs) {    
+  void run(double deltaMs) {    
     for (Point p : model.points) {
       color c = img.get((int)((p.x / model.xMax) * img.width), img.height - (int)((p.y / model.yMax) * img.height));
       colors[p.index] = color(hue(c) + colorMod.getValuef()%360, saturation(c), brightness(c) - ((p.fz - brightMod.getValuef())/p.fz));
@@ -60,7 +60,7 @@ class Spirality extends SCPattern {
     }
   }
     
-  public void run(int deltaMs) {
+  public void run(double deltaMs) {
     angle += deltaMs * 0.007;
     rad += deltaMs * .025 * direction;
     float x = model.xMax / 2 + cos(angle) * rad;
